Move createMessageRef() to TestingMessageRef
